Merge of leangrsi back to master

+++ libraries/TGRSIint/TGRSIint.cxx
@@ -84,13 +84,11 @@ void TGRSIint::ApplyOptions() {
    if(fAutoSort){
      TGRSILoop::Get()->SortMidas();
    }
-   bool foundCal = false;
-   if(fFragmentSort && TGRSIOptions::GetInputRoot().size()!=0)
-      TGRSIRootIO::Get()->MakeUserHistsFromFragmentTree();
+   //bool foundCal = false;
    if(TGRSIOptions::MakeAnalysisTree() && TGRSIOptions::GetInputRoot().size()!=0) { 
       TAnalysisTreeBuilder::Get()->StartMakeAnalysisTree();
    }
-   for(int x=0;x<TGRSIOptions::GetInputRoot().size();x++) {
+   for(int x=0;x<(int)(TGRSIOptions::GetInputRoot().size());x++) {
       //printf("TFile *_file%i = new TFile(\"%s\",\"read\")\n",x,TGRSIOptions::GetInputRoot().at(x).c_str());
       long error = ProcessLine(Form("TFile *_file%i = new TFile(\"%s\",\"read\");",x,TGRSIOptions::GetInputRoot().at(x).c_str()));
       if(error <=0) continue;
@@ -98,19 +96,30 @@ void TGRSIint::ApplyOptions() {
       printf("\tfile %s opened as _file%i\n",file->GetName(),x);
       TGRSIRootIO::Get()->LoadRootFile(file);
    }
-   if(TGRSIOptions::GetInputRoot().size() > 0) {
+   if(TGRSIOptions::GetInputRoot().size() > 0 && !fAutoSort && !fFragmentSort) {
       if(TGRSIOptions::GetInputRoot().at(0).find("fragment") != std::string::npos){
          Int_t chans_read = ProcessLine("TChannel::ReadCalFromTree(FragmentTree)");
          printf("Read calibration info for %d channels from \"%s\" FragmentTree\n",chans_read,TGRSIOptions::GetInputRoot().at(0).c_str()); 
+         TGRSIRunInfo::ReadInfoFromFile();
       }   
       if(TGRSIOptions::GetInputRoot().at(0).find("analysis") != std::string::npos){ 
          Int_t chans_read = ProcessLine("TChannel::ReadCalFromTree(AnalysisTree)");    
          printf("Read calibration info for %d channels from \"%s\" AnalysisTree\n",chans_read,TGRSIOptions::GetInputRoot().at(0).c_str());
+         TGRSIRunInfo::ReadInfoFromFile();
       }
    }
-  
-  if(TGRSIOptions::WorkHarder()) {
-      for(int x=0;x<TGRSIOptions::GetMacroFile().size();x++) {
+
+   if(TGRSIOptions::GetInputCal().size() > 0){
+		for(size_t i =0; i<TGRSIOptions::GetInputCal().size();++i){
+         TChannel::ReadCalFile(TGRSIOptions::GetInputCal().at(i).c_str());
+      }
+   }
+
+   if(fFragmentSort && TGRSIOptions::GetInputRoot().size()!=0)
+      TGRSIRootIO::Get()->MakeUserHistsFromFragmentTree();
+
+   if(TGRSIOptions::WorkHarder()) {
+      for(size_t x=0;x<TGRSIOptions::GetMacroFile().size();x++) {
          gROOT->Macro(TGRSIOptions::GetMacroFile().at(x).c_str());  
        // gROOT->ProcessLineSync(Form(".x %s",TGRSIOptions::GetMacroFile().at(x).c_str()));
       }
@@ -260,11 +269,11 @@ void TGRSIint::GetOptions(int *argc, char **argv) {
       } else if (sargv[0] == '-' && sargv[1] != '-') { //single char options.
         sargv = sargv.substr(1);  //drop the minus;
         int defaultcounter = 0;
-        for(int c=0;c<sargv.length();c++) {
+        for(size_t c=0;c<sargv.length();c++) {
           char key = sargv[c];
           switch(toupper(key)) {
             case 'A':
-      		  printf(DBLUE "Atempting to make analysis trees." RESET_COLOR "\n");
+      		  printf(DBLUE "Attempting to make analysis trees." RESET_COLOR "\n");
          	  TGRSIOptions::SetMakeAnalysisTree();
 		        break;
             case 'Q':
@@ -323,7 +332,7 @@ void TGRSIint::GetOptions(int *argc, char **argv) {
           TGRSIOptions::SetProgressDialog(false);
        } else if((temp.compare("bad_frags")==0)     || (temp.compare("write_bad_frags")==0) ||
                  (temp.compare("bad_fragments")==0) || (temp.compare("write_bad_fragments")==0)) {
-          printf(DBLUE "    failed fragements being written too BadFragmentTree." RESET_COLOR "\n");
+          printf(DBLUE "    failed fragements being written to BadFragmentTree." RESET_COLOR "\n");
           TGRSIOptions::SetWriteBadFrags(true);
        } else if(temp.compare("help")==0) {
           fPrintHelp = true;
@@ -331,11 +340,9 @@ void TGRSIint::GetOptions(int *argc, char **argv) {
           // useful when dealing with midas file that have corrupt odbs in them .
           TGRSIOptions::SetIgnoreFileOdb(true);          
        } else if(temp.compare("ignore_epics")==0) { 
-          // useful when dealing with midas file that have corrupt odbs in them .
           TGRSIOptions::SetIgnoreEpics(true);          
-       } else if(temp.compare("ignore_sclr")==0) { 
-          // useful when dealing with midas file that have corrupt odbs in them .
-          TGRSIOptions::SetIgnoreSCLR(true);          
+       } else if(temp.compare("ignore_scaler")==0) { 
+          TGRSIOptions::SetIgnoreScaler(true);          
        } else {
           printf(DBLUE  "    option: " DYELLOW "%s " DBLUE "passed but not understood." RESET_COLOR "\n",temp.c_str());
        }
@@ -350,7 +357,7 @@ void TGRSIint::GetOptions(int *argc, char **argv) {
                 pwd = argv[i]; 
                 argv[i]= null;
              //} 
-               printf("\tOption %s is a directory, ignoing for now.\n",pwd.c_str()); 
+               printf("\tOption %s is a directory, ignoring for now.\n",pwd.c_str());
           } else if (size > 0) {
              // if file add to list of files to be processed
              FileAutoDetect(argv[i],size);
@@ -359,8 +366,10 @@ void TGRSIint::GetOptions(int *argc, char **argv) {
              printf("file %s has size 0, skipping\n", dir);
           }
        } else {
-          //file does not exsist... assuming output file.
-          FileAutoDetect(argv[i],-1);
+          //file does not exist... complain to the user about this
+          if(!FileAutoDetect(argv[i],-1)) {
+            printf(DRED "File %s does not exist, ignoring it!" RESET_COLOR "\n",argv[i]);
+          }
           argv[i] = null;
        }   
     }
@@ -389,43 +398,40 @@ bool TGRSIint::FileAutoDetect(std::string filename, long filesize) {
    //first search for extensions.
    std::string ext = filename.substr(filename.find_last_of('.')+1);
    //printf("\text = %s\n",ext.c_str());
-   if(ext.compare("root")==0) {
+   if(ext.compare("root")==0 && filesize > 0) {
       //printf("\tFound root file: %s\n",filename.c_str());
       //fInputRootFile->push_back(filename);
       TGRSIOptions::AddInputRootFile(filename);
       return true;
-   } else if(ext.compare("mid")==0 || ext.compare("bz2")==0) {
+   } else if((ext.compare("mid")==0 || ext.compare("bz2")==0) && filesize > 0) {
       //printf("\tFound midas file: %s\n",filename.c_str());
       //fInputMidasFile->push_back(filename);
       TGRSIOptions::AddInputMidasFile(filename);
       fAutoSort = true;
       return true;
-   } else if(ext.compare("cal")==0) { 
+   } else if(ext.compare("cal")==0 && filesize > 0) { 
       //printf("\tFound custom calibration file: %s\n",filename.c_str());
       //fInputCalFile->push_back(filename);
       TGRSIOptions::AddInputCalFile(filename);
       return true;
-   } else if(ext.compare("info")==0) { 
-      if(!TGRSIOptions::ExternalRunInfo())
-        TGRSIOptions::AddExternalRunInfo(filename);
-      else
-        printf(DRED "more than one run info found!  discarding %s\n",filename.c_str());
-      //if(TGRSIRunInfo::ReadInfoFile(filename.c_str()))
-      //   return true;
-      //else {
-      //   printf("Problem reading run-info file %s\n",filename.c_str());
-      //   return false;
-      //}
-   } else if(ext.compare("xml")==0) { 
+   } else if(ext.compare("info")==0 && filesize > 0) { 
+      if(TGRSIRunInfo::ReadInfoFile(filename.c_str()))
+         return true;
+      else {
+         printf("Problem reading run-info file %s\n",filename.c_str());
+         return false;
+      }
+   } else if(ext.compare("xml")==0 && filesize > 0) { 
       //fInputOdbFile->push_back(filename);
       TGRSIOptions::AddInputOdbFile(filename);
       //printf("\tFound xml odb file: %s\n",filename.c_str());
       return true;
-   } else if(ext.compare("odb")==0) { 
+   } else if(ext.compare("odb")==0 && filesize > 0) { 
       //printf("\tFound c-like odb file: %s\n",filename.c_str());
       printf("c-like odb structures can't be read yet.\n");
       return false;
    } else if((ext.compare("c")==0) || (ext.compare("C")==0) || (ext.compare("c+")==0) || (ext.compare("C+")==0)) {
+      //scripts are the only files that don't have to exist, they may also be found in the macro-paths
       TGRSIOptions::AddMacroFile(filename);
       return true;
    } else {
@@ -437,9 +443,15 @@ bool TGRSIint::FileAutoDetect(std::string filename, long filesize) {
 
 
 bool TGRSIInterruptHandler::Notify() {
+   static int abort_times =0;
+   if(abort_times>3) {
+     printf("\n" DRED BG_WHITE  "   Control-c was pressed harder.   " RESET_COLOR  SHOW_CURSOR "\n");
+     abort();
+   }
    printf("\n" DRED BG_WHITE  "   Control-c was pressed.   " RESET_COLOR  SHOW_CURSOR "\n");
-   abort();
+   abort_times++;
    gApplication->Terminate();
+
    return true;
 }
